Moved StructArray to its own module.
